+++ src/api.ts
@@ -1,4 +1,3 @@
-export {};
 let XMLHttpRequest;
 if (typeof window === 'undefined') {
   XMLHttpRequest = require('xmlhttprequest').XMLHttpRequest;
@@ -29,10 +28,10 @@ const _HelperCallbacks = {
  * @property {boolean} success
  */
 class Message {
-	public http_code: any;
-	public message: any;
-	public success: any;
-	public error: any;
+	public http_code: number;
+	public message: string;
+	public success: boolean;
+	public error: string;
 
   constructor(http_code, message) {
     this.http_code = http_code;
@@ -50,7 +49,7 @@ class Message {
 /**
  * Handles Thor API requests
  */
-class API {
+export class API {
 	public host: any;
 	public token: any;
 	public error: any;
@@ -58,8 +57,8 @@ class API {
 	public version: any;
 	public success: any;
 	public status: any;
-	public http_code: any;
-	public id: any;
+	public http_code: number;
+	public id: string;
 
   /** @typedef Token
    * @property {string} expires Token expiration date
@@ -72,7 +71,7 @@ class API {
    * @param {string} config.host=https://api.smartslice.xyz API protocol and host name
    * @param {Token} config.token Authorization bearer token to use in API calls
    */
-  constructor(config) {
+  constructor(config?) {
     if (config === undefined) {
       config = {
         host: 'https://api.smartslice.xyz',
@@ -100,7 +99,7 @@ class API {
 
   _request(method, route, success, error, data?) {
     const xhttp = new XMLHttpRequest();
-    var response;
+    let response;
 
     xhttp.onreadystatechange = function() {
       if (xhttp.readyState === 4) {
@@ -806,5 +805,3 @@ class API {
     );
   }
 }
-
-module.exports = API;
