+++ src/fea/model/fracture.ts
@@ -1,0 +1,9 @@
+/**
+ * Fracture properties.
+ */
+ export interface Fracture {
+  /**
+   * Mode I fracture toughness.
+   */
+  KIc: number;
+ }
